Move the modification counters to indices package:
It is currently used with indexes only, hence belongs there.
